base step

+++ ProjectDash.Backend/ProjectDash.Application/Employees/Commands/DeleteEmployee/DeleteEmployeeCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using MediatR;
+using Microsoft.EntityFrameworkCore;
 using ProjectDash.Application.Common.Exceptions;
 using ProjectDash.Domain;
 using ProjectDash.Domain.Interfaces;
@@ -18,6 +19,13 @@ namespace ProjectDash.Application.Employees.Commands.DeleteEmployee
             if (entity == null)
             {
                 throw new NotFoundException(nameof(Employee), request.Id);
+            }
+            var projects = await _dbContext.Projects
+                .Where(project => project.ProjectLeaderId == request.Id)
+                .ToListAsync(cancellationToken);
+            if (projects.Count != 0)
+            {
+                throw new Exception("The Employee cannot be deleted if one's is a Project Leader.");
             }
 
             _dbContext.Employees.Remove(entity);
